MAJOR-main/MAJOR-main/Posture/Posture/requirements.txt
tensorflow
opencv-python
numpy
matplotlib
scikit-learn
pyttsx3
pillow
